不知道要怎麼做 auto generate class properties QQQ
