+++ src/resolvers/types.ts
@@ -1,19 +1,29 @@
-export enum Role {
-  User = 'USER',
-  Admin = 'ADMIN',
+import { FindOneOptions, getConnection } from 'typeorm';
+import { InputType, Field } from 'type-graphql';
+// import { InjectConnection } from 'typeorm-typedi-extensions';
+
+@InputType()
+export class FindOptions<Entity> implements Partial<FindOneOptions> {
+
+  @Field(() => String, { nullable: true })
+  public order?: {
+    [P in keyof Entity]?: 'ASC' | 'DESC';
+  };
+
+  @Field(() => String, { nullable: true })
+  public where?: {
+    [key: string]: any;
+  };
 }
 
-export type Context = {
-  me: {
-    id: string;
-    displayName: string;
-    email: string;
-    roles: Role[];
-  };
-  dataLoader: {
-    initialized: boolean;
-    loaders: {
-      [key: string]: any;
-    } & object;
-  }
-};
+// export async function createEntityFindOptions(entity: string) {
+
+//   const connection = await getConnection();
+//   const properties = Object.keys(connection.getMetadata(entity).propertiesMap);
+//   console.log(properties);
+//   return InputType(`${entity}Options`)(
+//     class {
+//       public a: string;
+//     },
+//   );
+// }
